backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
anthropic>=0.40.0
openai>=2.0.0
langfuse>=3.0,<4
sse-starlette==2.1.0
python-dotenv==1.0.1
